Started death and wave states
- Death screen
- "Rewind" on Restart
- Quit
- Wave count
- Enemy count
