+++ Assets/Resources/Player/Player.prefab
@@ -96,6 +96,7 @@ GameObject:
   - component: {fileID: 5841979810881545746}
   - component: {fileID: 5841979810881545757}
   - component: {fileID: 5841979810881545744}
+  - component: {fileID: 3783479438812845471}
   m_Layer: 0
   m_Name: Player
   m_TagString: Player
@@ -164,6 +165,18 @@ MonoBehaviour:
   currentHealth: 0
   maxHealth: 0
   owner: 0
+--- !u!114 &3783479438812845471
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5841979810881545756}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 99d5b172b643fc0439fcada75dde81b4, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &5841979812240164344
 GameObject:
   m_ObjectHideFlags: 0
