+++ Assets/Scripts/Enemy/EnemyBehaviour.cs
@@ -66,7 +66,7 @@ public class EnemyBehaviour : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        if (!_initialized) return;
+        if (!_initialized || TimeBody.isRewinding) return;
         
         // Look. Determine look direction.
         _direction = (_targetTransform.position - transform.position).normalized;
