+++ Assets/Scripts/TimeStuff/PointInTime.cs
@@ -1,0 +1,15 @@
+using UnityEngine;
+
+public class PointInTime
+{
+
+	public Vector3 position;
+	public Quaternion rotation;
+
+	public PointInTime(Vector3 _position, Quaternion _rotation)
+	{
+		position = _position;
+		rotation = _rotation;
+	}
+
+}
