+++ Assets/Scripts/TimeStuff/PointInTime.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fb4942e5f0f186246a8baec2fe759354
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
