+++ Assets/Scripts/UI/deathUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 69137c045574ea149ade1a236c3defa1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
